Support kitchen-dokken as :dokken_encrypt_databags
Allow use of kitchen-dokken by using the :dokken_encrypt_databags
provisioner key, with config like:

```
 platforms:
  - name: ubuntu-16.04
    driver: { name: dokken }
    provisioner: { name: dokken_encrypt_databags }
```

Introduces `Kitchen::Provisioner::EncryptDatabags` that includes the
common functionality required to decrypt data bags on the way into a
node's sandbox. This should make it easy to extend other provisioners to
support this functionality in future.
