+++ .kitchen.yml
@@ -1,16 +1,23 @@
 ---
-driver:
-  name: docker
 
 provisioner:
-  name: chef_zero_encrypt_databags
   data_bags_path: spec/fixtures/data_bags
   encrypted_data_bag_secret_key_path: spec/fixtures/encrypted_data_bag_secret
 
 platforms:
-  - name: ubuntu-16.04
+  - name: ubuntu
 
 suites:
-  - name: default
-    attributes: {}
+  - name: chef-zero
     run_list: fake::default
+    driver: { name: docker }
+    provisioner: { name: chef_zero_encrypt_databags }
+  - name: dokken
+    run_list: fake::default
+    driver: { name: dokken, image: ubuntu }
+    transport: { name: dokken }
+    provisioner: { name: dokken_encrypt_databags }
+    verifier: {
+      root_path: /opt/verifier,
+      sudo: false
+    }
