+++ .ruby-version
@@ -1,1 +1,1 @@
-2.3.1
+2.4.1
