+++ Berksfile
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 source "https://supermarket.chef.io"
 
 cookbook "fake", path: "test/fixtures/cookbooks/fake"
