+++ kitchen-encrypt-databags.gemspec
@@ -1,13 +1,14 @@
 # coding: utf-8
 # frozen_string_literal: true
+
 Gem::Specification.new do |gem|
   gem.name          = "kitchen-encrypt-databags"
-  gem.version       = "0.1.1"
+  gem.version       = "0.1.2"
   gem.summary       = "automatically encrypt databags when provisioning sandbox"
   gem.description   = "test-kitchen add-on"
   gem.authors       = ["Lawrence Jones"]
   gem.homepage      = "https://github.com/lawrencejones/kitchen-encrypt-databags"
-  gem.email         = %w(lawrjone@gmail.com)
+  gem.email         = %w[lawrjone@gmail.com]
   gem.license       = "MIT"
 
   gem.files         = Dir["lib/**/*.rb"]
