+++ lib/kitchen/provisioner/dokken_encrypt_databags.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 require "kitchen"
-require "kitchen/provisioner/chef_zero"
+require "kitchen/provisioner/dokken"
 
 require_relative "encrypt_databags"
 
 module Kitchen
   module Provisioner
-    class ChefZeroEncryptDatabags < ::Kitchen::Provisioner::ChefZero
+    class DokkenEncryptDatabags < ::Kitchen::Provisioner::Dokken
       include EncryptDatabags
     end
   end
