+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 $LOAD_PATH << "../lib"
 
 RSpec.configure do |config|
