+++ test/fixtures/cookbooks/fake/metadata.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
+
 name "fake"
 version "1.0.0"
