+++ test/fixtures/cookbooks/fake/recipes/default.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # This recipe isn't meant to do anything sensible- just to test that password inside the
 # sudoers_enc databag is accessible by decrypting the JSON with the secret key provisioned
 # on the sandbox.
