--- test/integration/default/serverspec/spec_helper.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-require "serverspec"
-set :backend, :exec
