+++ test/integration/docker/serverspec/root_password_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
-require "spec_helper"
+
+require "serverspec"
+set :backend, :exec
 
 describe file("/etc/root_password") do
   it { should contain("much-entropy") }
