+++ test/integration/dokken/serverspec/root_password_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "serverspec"
+set :backend, :exec
+
+describe file("/etc/root_password") do
+  it { should contain("much-entropy") }
+end
